+++ tests/CalibreIcuTest.php
@@ -5,9 +5,10 @@
 
 use BicBucStriim\Calibre\Calibre;
 use BicBucStriim\Calibre\CalibreFilter;
+use PHPUnit\Framework\Attributes\CoversClass;
 
-#[\PHPUnit\Framework\Attributes\CoversClass(\BicBucStriim\Calibre\Calibre::class)]
-#[\PHPUnit\Framework\Attributes\CoversClass(\BicBucStriim\Calibre\CalibreFilter::class)]
+#[CoversClass(Calibre::class)]
+#[CoversClass(CalibreFilter::class)]
 class CalibreIcuTest extends PHPUnit\Framework\TestCase
 {
     public const CDB4 = './tests/fixtures/lib4/metadata.db';